Make the code more portable against library function deficiencies
On Windows, vsnprintf returns the wrong value, and printf/scanf don't
understand long longs. This works around the problem via allowing for
unreasonable vsnprintf return values, and via using int_least64_t
instead of long long for printf/scanf.

This contains the same fixes as 704a2381, except that they're written
properly rather than being incredibly hacky.

+++ libnethack/util/makedefs.c
@@ -368,7 +368,7 @@ do_date(const char *outfile, int printdates)
     *c = '\0';  /* strip off the '\n' */
     if (printdates) {
         fprintf(ofp, "#define BUILD_DATE \"%s\"\n", cbuf);
-        fprintf(ofp, "#define BUILD_TIME (%ldL)\n", clocktim);
+        fprintf(ofp, "#define BUILD_TIME (%ldL)\n", (long)clocktim);
         fprintf(ofp, "\n");
     }
